feat: make navbar view template
