+++ frontend/src/Navbar/firstRow/NavbarProfileImage.tsx
