+++ frontend/src/Navbar/firstRow/navbarFirstRow.scss
@@ -24,7 +24,7 @@ button {
     cursor: pointer;
 
     & .navbarFirstRow__upLoadBtn {
-        font-size: 0.7rem;
+        font-size: 0.75rem;
         font-weight: 700;
       }
