+++ frontend/src/Navbar/firstRow/navbarProfileImage.scss
