+++ frontend/src/Navbar/secondRow/NavbarSecondRow.tsx
@@ -1,0 +1,15 @@
+import React from 'react';
+import './navbarSecondRow.scss';
+import CategoryArea from './CategoryArea';
+import SearchArea from './SearchArea';
+
+function NavbarSecondRow() {
+  return (
+    <section className="navbarSecondRow">
+      <CategoryArea />
+      <SearchArea/>
+    </section>
+  );
+}
+
+export default NavbarSecondRow;
