+++ frontend/src/Navbar/secondRow/navbarSecondRow.scss
@@ -1,5 +1,4 @@
 $navbar-height: 14vh;
-$common-border-radius: 5px;
 
 // common
 button {
@@ -13,14 +12,7 @@ button {
   width: 100%;
   height: 40%;
   display: flex;
+  justify-content: space-between;
   align-items: center;
-
-  & .categoryArea {
-
-  }
-
-  & .searchArea {
-
-  }
 }
