+++ frontend/src/Navbar/secondRow/searchArea.scss
@@ -1,0 +1,24 @@
+// variable
+$common-border-radius: 5px;
+
+// navbar search Area
+.searchArea {
+    width: 18%;
+    height: 70%;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    background-color: #f3f3f3;
+    border-radius: $common-border-radius;
+
+    & input {
+      background-color: rgba(0,0,0,0);
+      font-size: 0.7rem;
+      font-weight: 500;
+      border: none;
+
+      &:focus {
+        outline: none;
+      }
+    }
+  }
